requirements.txt
flask>=3.0
twilio>=9.0
google-api-python-client>=2.120
google-auth>=2.28
google-auth-httplib2>=0.2
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.0
